fixing previous fix.
The variable newrow is depreciated and no longer used. I have added a catch that will look for all whitespace in the input file and give a warning message automatically. This has been implemented for both the tree and msn. This should ensure faster loading of the page since it doesn't have to attempt to collect the data before loading.